Update the Unit test examples in the cool-core and Update the load fucntion for the parquet format data
