+++ cool-core/src/main/java/com/nus/cool/loader/CohortLoader.java
@@ -26,7 +26,6 @@ import com.nus.cool.core.cohort.CohortAggregation;
 import com.nus.cool.core.cohort.CohortKey;
 import com.nus.cool.core.cohort.CohortQuery;
 import com.nus.cool.core.cohort.CohortSelection;
-import com.nus.cool.core.cohort.FieldSet;
 import com.nus.cool.core.cohort.QueryResult;
 import com.nus.cool.core.io.compression.SimpleBitSetCompressor;
 import com.nus.cool.core.io.readstore.ChunkRS;
@@ -38,13 +37,12 @@ import com.nus.cool.core.schema.TableSchema;
 import com.nus.cool.core.util.IntegerUtil;
 import com.nus.cool.core.util.converter.DayIntConverter;
 import com.nus.cool.core.util.converter.NumericConverter;
+import com.nus.cool.model.CoolModel;
+
 import java.io.DataOutputStream;
 import java.io.File;
-import java.io.FileFilter;
 import java.io.FileOutputStream;
-import java.io.FilenameFilter;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.List;
 import java.util.Map;
