+++ cool-core/src/main/java/com/nus/cool/model/CoolModel.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.nus.cool.loader;
+package com.nus.cool.model;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -26,9 +26,12 @@ import com.nus.cool.core.io.storevector.InputVector;
 import com.nus.cool.core.io.readstore.CubeRS;
 import com.nus.cool.core.io.readstore.CohortRS;
 import com.nus.cool.core.schema.TableSchema;
+import com.nus.cool.core.util.config.DataLoaderConfig;
+import com.nus.cool.loader.DataLoader;
 
 import java.io.*;
 import java.nio.ByteOrder;
+import java.nio.file.StandardCopyOption;
 import java.util.*;
 
 /**
