+++ cool-queryserver/src/main/java/com.nus.cool.queryserver/QueryServerModel.java
@@ -6,7 +6,7 @@ import static com.google.common.base.Preconditions.checkArgument;
 import com.nus.cool.core.cohort.ExtendedCohortQuery;
 import com.nus.cool.core.io.readstore.CubeRS;
 import com.nus.cool.core.io.storevector.InputVector;
-import com.nus.cool.loader.CoolModel;
+import com.nus.cool.model.CoolModel;
 import com.nus.cool.loader.ExtendedResultTuple;
 
 import javax.ws.rs.core.Response;
