+++ pom.xml
@@ -16,8 +16,8 @@
     <modules>
         <module>cool-core</module>
         <module>cool-extensions</module>
-        <module>cool-examples</module>
-        <module>cool-queryserver</module>
+<!--        <module>cool-examples</module>-->
+<!--        <module>cool-queryserver</module>-->
     </modules>
 
 
@@ -75,13 +75,6 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                    <excludes>
-                        <exclude>
-                            **/*IntegrationTest.java
-                        </exclude>
-                    </excludes>
-                </configuration>
             </plugin>
 
             <plugin>
